[Bronze V] Title: 이상한 기호, Time: 48 ms, Memory: 5496 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -1,4 +1,4 @@
             int[] n = Array.ConvertAll(Console.ReadLine().Split(), int.Parse);
-            long a = (n[0] + n[1]) * (n[0] - n[1]);
+            long a = (long)(n[0] + n[1]) * (n[0] - n[1]);
 
             Console.WriteLine(a);
